Test part 1

+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: oelkhiar <oelkhiar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/17 11:14:51 by oelkhiar          #+#    #+#             */
-/*   Updated: 2022/10/18 17:19:14 by oelkhiar         ###   ########.fr       */
+/*   Updated: 2022/10/18 21:24:40 by oelkhiar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_memmove(void *dst, const void *src, size_t len)
 {
 	unsigned char	*d;
 	unsigned char	*s;
-	int	i;
+	size_t			i;
 
 	d = (unsigned char *)dst;
 	s = (unsigned char *)src;
@@ -36,11 +36,3 @@ void	*ft_memmove(void *dst, const void *src, size_t len)
 	}
 	return (dst);
 }
-// int main(void)
-// {
-// 	char str2[] = "mahmudul hasan";
-// 	char *ret;
-// 	ret = ft_memmove(str2, str2 + 5, 10);  fix overlap
-// 	printf("%s", ret); => dul hasan
-// 	return (0);
-// }
